+++ src/App.vue
@@ -1,6 +1,6 @@
 <template>
   <v-app>
-    <v-app-bar app color="grey-darken-4" dark>
+    <v-app-bar app color="grey-darken-4" dark class="border-bottom-gray">
       <v-app-bar-nav-icon @click="appStore.toggleMainNav" />
       <v-toolbar-title>Vuetify NFT Gallery (TS)</v-toolbar-title>
       <v-spacer></v-spacer>
@@ -16,14 +16,13 @@
       v-model="appStore.isMainNavShowing"
       color="grey-darken-4"
     >
-      <v-list>
-        <v-divider></v-divider>
-
+      <v-list class="py-0">
         <v-list-item link :to="'/'">
           <v-list-item-title
             ><v-icon class="mr-2">mdi-home</v-icon>Home</v-list-item-title
           >
         </v-list-item>
+        <v-divider></v-divider>
 
         <v-list-item link :to="'/useEvmNftGallery'">
           <v-list-item-title>
@@ -31,20 +30,25 @@
             >useEvmNftGallery</v-list-item-title
           >
         </v-list-item>
+        <v-divider></v-divider>
         <v-list-item link :to="'/useEvmMetaDataGallery'">
           <v-list-item-title>
             <v-icon class="mr-2">mdi-image-multiple</v-icon
             >useEvmMetaDataGallery</v-list-item-title
           >
         </v-list-item>
+        <v-divider></v-divider>
       </v-list>
     </v-navigation-drawer>
     <v-main>
       <router-view />
+      <AppFooter />
     </v-main>
   </v-app>
 </template>
 
 <script lang="ts" setup>
+import AppFooter from './components/AppFooter.vue';
+
 const appStore = useAppStore();
 </script>
